Исправлен shebang скрипта
