+++ Dockerfile
@@ -31,10 +31,11 @@ COPY . /app
 
 # 8. Копируем скрипт миграции и делаем его исполняемым
 COPY entrypoint.sh /app/entrypoint.sh
-RUN chmod +x /app/entrypoint.sh
+RUN sed -i 's/\r$//' /app/entrypoint.sh \
+    && chmod +x /app/entrypoint.sh
 
 # 9. Скрипт-мигратор запускается первым, затем передаёт управление UV
 ENTRYPOINT ["/app/entrypoint.sh"]
 
 # 6. Команда по умолчанию: запускаем через UV (он найдёт ваши зависимости и сам выполнит main.py)
-CMD ["uv", "run", "main.py"]
+CMD ["uv", "run", "app/main.py"]
